==> Spatial_transcriptomics/03-merge_adata.ipynb <==
import os
import scanpy as sc
import pandas as pd
import numpy as np

adata_dir = 'adata'
samples = [f[5:-5] for f in os.listdir(adata_dir)] 

genes = pd.read_csv('cnmf_genes.csv', header=None, index_col=0).index

adatas = []
for sample in samples:
    _adata = sc.read(os.path.join(adata_dir,'adata%s.h5ad'%sample))
    _adata.obs['patient'] = np.repeat(sample, _adata.n_obs)
    
    use_genes = _adata.var.index.intersection(genes)
    _adata = _adata[:,use_genes]
    
    adatas.append(_adata)

adata = adatas[0]
if len(adata)>1:
    adata = adata.concatenate(adatas[1::],batch_key='batch_number', join='outer', fill_value=0)

# correct a few columns in var
vcol = adata.var.columns
cols_to_change = ['gene_ids','genome','mt','feature_types']
for newcol in cols_to_change:
    coi = [c for c in vcol if newcol in c]
    adata.var[newcol] = adata.var[coi].fillna(method='bfill', axis=1).iloc[:, 0]
    adata.var.drop(coi, axis=1, inplace=True)

bool_col = 'highly_variable'
coi = [c for c in vcol if bool_col in c]
adata.var[coi] = adata.var[coi].fillna(False)

adata.write('adata.h5ad')

adata
